# ShunFengParcel/api/helloworld/v1/kuai_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v3.21.11
// source: api/helloworld/v1/kuai.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationKuaiCancelOrder = "/helloworld.v1.Kuai/CancelOrder"
const OperationKuaiCreateOrder = "/helloworld.v1.Kuai/CreateOrder"
const OperationKuaiGetCourierPerformance = "/helloworld.v1.Kuai/GetCourierPerformance"
const OperationKuaiGetTask = "/helloworld.v1.Kuai/GetTask"
const OperationKuaiHandleException = "/helloworld.v1.Kuai/HandleException"
const OperationKuaiIncomeList = "/helloworld.v1.Kuai/IncomeList"
const OperationKuaiLogin = "/helloworld.v1.Kuai/Login"
const OperationKuaiOrderDetail = "/helloworld.v1.Kuai/OrderDetail"
const OperationKuaiOrderList = "/helloworld.v1.Kuai/OrderList"
const OperationKuaiPerformance = "/helloworld.v1.Kuai/Performance"
const OperationKuaiReassignOrder = "/helloworld.v1.Kuai/ReassignOrder"
const OperationKuaiRegister = "/helloworld.v1.Kuai/Register"
const OperationKuaiStuUpd = "/helloworld.v1.Kuai/StuUpd"
const OperationKuaiTakeTask = "/helloworld.v1.Kuai/TakeTask"
const OperationKuaiTaskList = "/helloworld.v1.Kuai/TaskList"

type KuaiHTTPServer interface {
	// CancelOrder 订单取消
	CancelOrder(context.Context, *CancelOrderRequest) (*CancelOrderReply, error)
	// CreateOrder生成订单
	CreateOrder(context.Context, *CreateOrderRequest) (*CreateOrderReply, error)
	// GetCourierPerformance数据统计
	GetCourierPerformance(context.Context, *GetCourierPerformanceRequest) (*GetCourierPerformanceReply, error)
	// GetTask个人任务详情
	GetTask(context.Context, *GetTaskRequest) (*GetTaskReply, error)
	// HandleException异常处理接口
	HandleException(context.Context, *HandleExceptionRequest) (*HandleExceptionReply, error)
	// IncomeList 收入明细列表
	IncomeList(context.Context, *IncomeListRequest) (*IncomeListReply, error)
	Login(context.Context, *LoginRequest) (*LoginReply, error)
	// OrderDetail 订单详情（运营与客服深度查询），缓存优先读库回源
	OrderDetail(context.Context, *OrderDetailRequest) (*OrderDetailReply, error)
	// OrderList订单列表
	OrderList(context.Context, *OrderListRequest) (*OrderListReply, error)
	// Performance绩效排行榜
	Performance(context.Context, *PerformanceRequest) (*PerformanceReply, error)
	// ReassignOrder 订单改派
	ReassignOrder(context.Context, *ReassignOrderRequest) (*ReassignOrderReply, error)
	Register(context.Context, *RegisterRequest) (*RegisterReply, error)
	// StuUpd修改快递员在线状态
	StuUpd(context.Context, *StuUpdRequest) (*StuUpdReply, error)
	// TakeTask快递员点击接单
	TakeTask(context.Context, *TakeTaskRequest) (*TakeTaskReply, error)
	// TaskList任务展示
	TaskList(context.Context, *TaskListRequest) (*TaskListReply, error)
}

func RegisterKuaiHTTPServer(s *http.Server, srv KuaiHTTPServer) {
	r := s.Route("/")
	r.POST("/gettask", _Kuai_GetTask0_HTTP_Handler(srv))
	r.POST("/taskList", _Kuai_TaskList0_HTTP_Handler(srv))
	r.POST("/login", _Kuai_Login0_HTTP_Handler(srv))
	r.POST("/register", _Kuai_Register0_HTTP_Handler(srv))
	r.POST("/stuUpd", _Kuai_StuUpd0_HTTP_Handler(srv))
	r.POST("/addTask", _Kuai_TakeTask0_HTTP_Handler(srv))
	r.POST("/createOrder", _Kuai_CreateOrder0_HTTP_Handler(srv))
	r.POST("/performance", _Kuai_Performance0_HTTP_Handler(srv))
	r.POST("/orders", _Kuai_OrderList0_HTTP_Handler(srv))
	r.POST("/income/list", _Kuai_IncomeList0_HTTP_Handler(srv))
	r.POST("/handleException", _Kuai_HandleException0_HTTP_Handler(srv))
	r.POST("/orders/cancel", _Kuai_CancelOrder0_HTTP_Handler(srv))
	r.POST("/orders/reassign", _Kuai_ReassignOrder0_HTTP_Handler(srv))
	r.POST("/order/detail", _Kuai_OrderDetail0_HTTP_Handler(srv))
	r.POST("/count/performance", _Kuai_GetCourierPerformance0_HTTP_Handler(srv))
}

func _Kuai_GetTask0_HTTP_Handler(srv KuaiHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetTaskRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationKuaiGetTask)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetTask(ctx, req.(*GetTaskRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetTaskReply)
		return ctx.Result(200, reply)
	}
}

func _Kuai_TaskList0_HTTP_Handler(srv KuaiHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in TaskListRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationKuaiTaskList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.TaskList(ctx, req.(*TaskListRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*TaskListReply)
		return ctx.Result(200, reply)
	}
}

func _Kuai_Login0_HTTP_Handler(srv KuaiHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in LoginRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationKuaiLogin)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Login(ctx, req.(*LoginRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*LoginReply)
		return ctx.Result(200, reply)
	}
}

func _Kuai_Register0_HTTP_Handler(srv KuaiHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in RegisterRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationKuaiRegister)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Register(ctx, req.(*RegisterRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*RegisterReply)
		return ctx.Result(200, reply)
	}
}

func _Kuai_StuUpd0_HTTP_Handler(srv KuaiHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in StuUpdRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationKuaiStuUpd)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.StuUpd(ctx, req.(*StuUpdRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*StuUpdReply)
		return ctx.Result(200, reply)
	}
}

func _Kuai_TakeTask0_HTTP_Handler(srv KuaiHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in TakeTaskRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationKuaiTakeTask)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.TakeTask(ctx, req.(*TakeTaskRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*TakeTaskReply)
		return ctx.Result(200, reply)
	}
}

func _Kuai_CreateOrder0_HTTP_Handler(srv KuaiHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateOrderRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationKuaiCreateOrder)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateOrder(ctx, req.(*CreateOrderRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateOrderReply)
		return ctx.Result(200, reply)
	}
}

func _Kuai_Performance0_HTTP_Handler(srv KuaiHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in PerformanceRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationKuaiPerformance)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Performance(ctx, req.(*PerformanceRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*PerformanceReply)
		return ctx.Result(200, reply)
	}
}

func _Kuai_OrderList0_HTTP_Handler(srv KuaiHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in OrderListRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationKuaiOrderList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.OrderList(ctx, req.(*OrderListRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*OrderListReply)
		return ctx.Result(200, reply)
	}
}

func _Kuai_IncomeList0_HTTP_Handler(srv KuaiHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in IncomeListRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationKuaiIncomeList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.IncomeList(ctx, req.(*IncomeListRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*IncomeListReply)
		return ctx.Result(200, reply)
	}
}

func _Kuai_HandleException0_HTTP_Handler(srv KuaiHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in HandleExceptionRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationKuaiHandleException)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.HandleException(ctx, req.(*HandleExceptionRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*HandleExceptionReply)
		return ctx.Result(200, reply)
	}
}

func _Kuai_CancelOrder0_HTTP_Handler(srv KuaiHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CancelOrderRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationKuaiCancelOrder)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CancelOrder(ctx, req.(*CancelOrderRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CancelOrderReply)
		return ctx.Result(200, reply)
	}
}

func _Kuai_ReassignOrder0_HTTP_Handler(srv KuaiHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ReassignOrderRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationKuaiReassignOrder)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ReassignOrder(ctx, req.(*ReassignOrderRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ReassignOrderReply)
		return ctx.Result(200, reply)
	}
}

func _Kuai_OrderDetail0_HTTP_Handler(srv KuaiHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in OrderDetailRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationKuaiOrderDetail)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.OrderDetail(ctx, req.(*OrderDetailRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*OrderDetailReply)
		return ctx.Result(200, reply)
	}
}

func _Kuai_GetCourierPerformance0_HTTP_Handler(srv KuaiHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetCourierPerformanceRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationKuaiGetCourierPerformance)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetCourierPerformance(ctx, req.(*GetCourierPerformanceRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetCourierPerformanceReply)
		return ctx.Result(200, reply)
	}
}

type KuaiHTTPClient interface {
	// CancelOrder 订单取消
	CancelOrder(ctx context.Context, req *CancelOrderRequest, opts ...http.CallOption) (rsp *CancelOrderReply, err error)
	// CreateOrder生成订单
	CreateOrder(ctx context.Context, req *CreateOrderRequest, opts ...http.CallOption) (rsp *CreateOrderReply, err error)
	// GetCourierPerformance数据统计
	GetCourierPerformance(ctx context.Context, req *GetCourierPerformanceRequest, opts ...http.CallOption) (rsp *GetCourierPerformanceReply, err error)
	// GetTask个人任务详情
	GetTask(ctx context.Context, req *GetTaskRequest, opts ...http.CallOption) (rsp *GetTaskReply, err error)
	// HandleException异常处理接口
	HandleException(ctx context.Context, req *HandleExceptionRequest, opts ...http.CallOption) (rsp *HandleExceptionReply, err error)
	// IncomeList 收入明细列表
	IncomeList(ctx context.Context, req *IncomeListRequest, opts ...http.CallOption) (rsp *IncomeListReply, err error)
	Login(ctx context.Context, req *LoginRequest, opts ...http.CallOption) (rsp *LoginReply, err error)
	// OrderDetail 订单详情（运营与客服深度查询），缓存优先读库回源
	OrderDetail(ctx context.Context, req *OrderDetailRequest, opts ...http.CallOption) (rsp *OrderDetailReply, err error)
	// OrderList订单列表
	OrderList(ctx context.Context, req *OrderListRequest, opts ...http.CallOption) (rsp *OrderListReply, err error)
	// Performance绩效排行榜
	Performance(ctx context.Context, req *PerformanceRequest, opts ...http.CallOption) (rsp *PerformanceReply, err error)
	// ReassignOrder 订单改派
	ReassignOrder(ctx context.Context, req *ReassignOrderRequest, opts ...http.CallOption) (rsp *ReassignOrderReply, err error)
	Register(ctx context.Context, req *RegisterRequest, opts ...http.CallOption) (rsp *RegisterReply, err error)
	// StuUpd修改快递员在线状态
	StuUpd(ctx context.Context, req *StuUpdRequest, opts ...http.CallOption) (rsp *StuUpdReply, err error)
	// TakeTask快递员点击接单
	TakeTask(ctx context.Context, req *TakeTaskRequest, opts ...http.CallOption) (rsp *TakeTaskReply, err error)
	// TaskList任务展示
	TaskList(ctx context.Context, req *TaskListRequest, opts ...http.CallOption) (rsp *TaskListReply, err error)
}

type KuaiHTTPClientImpl struct {
	cc *http.Client
}

func NewKuaiHTTPClient(client *http.Client) KuaiHTTPClient {
	return &KuaiHTTPClientImpl{client}
}

// CancelOrder 订单取消
func (c *KuaiHTTPClientImpl) CancelOrder(ctx context.Context, in *CancelOrderRequest, opts ...http.CallOption) (*CancelOrderReply, error) {
	var out CancelOrderReply
	pattern := "/orders/cancel"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationKuaiCancelOrder))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// CreateOrder生成订单
func (c *KuaiHTTPClientImpl) CreateOrder(ctx context.Context, in *CreateOrderRequest, opts ...http.CallOption) (*CreateOrderReply, error) {
	var out CreateOrderReply
	pattern := "/createOrder"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationKuaiCreateOrder))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetCourierPerformance数据统计
func (c *KuaiHTTPClientImpl) GetCourierPerformance(ctx context.Context, in *GetCourierPerformanceRequest, opts ...http.CallOption) (*GetCourierPerformanceReply, error) {
	var out GetCourierPerformanceReply
	pattern := "/count/performance"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationKuaiGetCourierPerformance))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetTask个人任务详情
func (c *KuaiHTTPClientImpl) GetTask(ctx context.Context, in *GetTaskRequest, opts ...http.CallOption) (*GetTaskReply, error) {
	var out GetTaskReply
	pattern := "/gettask"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationKuaiGetTask))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// HandleException异常处理接口
func (c *KuaiHTTPClientImpl) HandleException(ctx context.Context, in *HandleExceptionRequest, opts ...http.CallOption) (*HandleExceptionReply, error) {
	var out HandleExceptionReply
	pattern := "/handleException"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationKuaiHandleException))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// IncomeList 收入明细列表
func (c *KuaiHTTPClientImpl) IncomeList(ctx context.Context, in *IncomeListRequest, opts ...http.CallOption) (*IncomeListReply, error) {
	var out IncomeListReply
	pattern := "/income/list"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationKuaiIncomeList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *KuaiHTTPClientImpl) Login(ctx context.Context, in *LoginRequest, opts ...http.CallOption) (*LoginReply, error) {
	var out LoginReply
	pattern := "/login"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationKuaiLogin))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// OrderDetail 订单详情（运营与客服深度查询），缓存优先读库回源
func (c *KuaiHTTPClientImpl) OrderDetail(ctx context.Context, in *OrderDetailRequest, opts ...http.CallOption) (*OrderDetailReply, error) {
	var out OrderDetailReply
	pattern := "/order/detail"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationKuaiOrderDetail))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// OrderList订单列表
func (c *KuaiHTTPClientImpl) OrderList(ctx context.Context, in *OrderListRequest, opts ...http.CallOption) (*OrderListReply, error) {
	var out OrderListReply
	pattern := "/orders"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationKuaiOrderList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Performance绩效排行榜
func (c *KuaiHTTPClientImpl) Performance(ctx context.Context, in *PerformanceRequest, opts ...http.CallOption) (*PerformanceReply, error) {
	var out PerformanceReply
	pattern := "/performance"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationKuaiPerformance))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ReassignOrder 订单改派
func (c *KuaiHTTPClientImpl) ReassignOrder(ctx context.Context, in *ReassignOrderRequest, opts ...http.CallOption) (*ReassignOrderReply, error) {
	var out ReassignOrderReply
	pattern := "/orders/reassign"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationKuaiReassignOrder))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *KuaiHTTPClientImpl) Register(ctx context.Context, in *RegisterRequest, opts ...http.CallOption) (*RegisterReply, error) {
	var out RegisterReply
	pattern := "/register"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationKuaiRegister))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// StuUpd修改快递员在线状态
func (c *KuaiHTTPClientImpl) StuUpd(ctx context.Context, in *StuUpdRequest, opts ...http.CallOption) (*StuUpdReply, error) {
	var out StuUpdReply
	pattern := "/stuUpd"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationKuaiStuUpd))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// TakeTask快递员点击接单
func (c *KuaiHTTPClientImpl) TakeTask(ctx context.Context, in *TakeTaskRequest, opts ...http.CallOption) (*TakeTaskReply, error) {
	var out TakeTaskReply
	pattern := "/addTask"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationKuaiTakeTask))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// TaskList任务展示
func (c *KuaiHTTPClientImpl) TaskList(ctx context.Context, in *TaskListRequest, opts ...http.CallOption) (*TaskListReply, error) {
	var out TaskListReply
	pattern := "/taskList"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationKuaiTaskList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}