# ShunFengParcel/openapi.yaml
...
openapi: 3.0.3
info:
    title: Kuai API
    version: 0.0.1
paths:
    /addTask:
        post:
            tags:
                - Kuai
            operationId: Kuai_TakeTask
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TakeTaskRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TakeTaskReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /count/performance:
        post:
            tags:
                - Kuai
            description: 数据统计
            operationId: Kuai_GetCourierPerformance
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GetCourierPerformanceRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetCourierPerformanceReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /createOrder:
        post:
            tags:
                - Kuai
            operationId: Kuai_CreateOrder
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateOrderRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateOrderReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /gettask:
        post:
            tags:
                - Kuai
            operationId: Kuai_GetTask
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GetTaskRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetTaskReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /handleException:
        post:
            tags:
                - Kuai
            operationId: Kuai_HandleException
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/HandleExceptionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HandleExceptionReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /income/list:
        post:
            tags:
                - Kuai
            operationId: Kuai_IncomeList
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/IncomeListRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/IncomeListReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /login:
        post:
            tags:
                - Kuai
            operationId: Kuai_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LoginReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /order/detail:
        post:
            tags:
                - Kuai
            description: 订单详情（运营与客服深度查询），缓存优先读库回源
            operationId: Kuai_OrderDetail
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/OrderDetailRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/OrderDetailReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /orders:
        post:
            tags:
                - Kuai
            operationId: Kuai_OrderList
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/OrderListRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/OrderListReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /orders/cancel:
        post:
            tags:
                - Kuai
            description: 订单取消
            operationId: Kuai_CancelOrder
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CancelOrderRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CancelOrderReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /orders/reassign:
        post:
            tags:
                - Kuai
            description: 订单改派
            operationId: Kuai_ReassignOrder
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ReassignOrderRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ReassignOrderReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /performance:
        post:
            tags:
                - Kuai
            operationId: Kuai_Performance
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PerformanceRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PerformanceReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /register:
        post:
            tags:
                - Kuai
            operationId: Kuai_Register
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RegisterRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/RegisterReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
...
        CreateOrderReply:
            type: object
            properties:
                orderId:
                    type: integer
                    format: int32
...
        GetCourierPerformanceRequest:
            type: object
            properties: {}
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...
        TakeTaskReply:
            type: object
            properties:
                task:
                    $ref: '#/components/schemas/Task'
        TakeTaskRequest:
            type: object
            properties:
                courierId:
                    type: string
...